test: LottoGenerator 테스트 추가
