+++ src/test/java/lotto/LottoShopTest.java
@@ -1,9 +1,6 @@
-package lotto.view;
+package lotto;
 
-import lotto.Lotto;
-import lotto.LottoGenerator;
-import lotto.LottoNumber;
-import lotto.LottoShop;
+import lotto.LottoGeneratorTest.TestLottoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -13,20 +10,6 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class LottoShopTest {
-
-  public static class TestLottoGenerator implements LottoGenerator {
-    @Override
-    public Lotto generate() {
-      return new Lotto(List.of(
-              new LottoNumber(1),
-              new LottoNumber(45),
-              new LottoNumber(2),
-              new LottoNumber(44),
-              new LottoNumber(3),
-              new LottoNumber(43)
-      ));
-    }
-  }
 
   @DisplayName("금액을 지불하면 가격에 맞는 로또 티켓들을 생성해주는 기능을 테스트한다.")
   @Test
